Desktop/hackathon-5/production/requirements.txt
# ── Web framework ────────────────────────────────────────────────────────────
fastapi==0.109.0
uvicorn[standard]==0.27.0

# ── AI / Agent SDK ───────────────────────────────────────────────────────────
openai>=1.75.0
openai-agents>=0.13.0

# ── Data validation ──────────────────────────────────────────────────────────
pydantic==2.5.3

# ── Database ─────────────────────────────────────────────────────────────────
asyncpg==0.29.0
pgvector==0.2.4

# ── Messaging ────────────────────────────────────────────────────────────────
aiokafka==0.10.0

# ── Logging ──────────────────────────────────────────────────────────────────
structlog==24.1.0

# ── Config ───────────────────────────────────────────────────────────────────
python-dotenv==1.0.0

# ── HTTP client ──────────────────────────────────────────────────────────────
httpx==0.26.0

# ── Channel integrations ─────────────────────────────────────────────────────
twilio==8.10.0
google-api-python-client==2.111.0
google-auth==2.27.0
google-auth-httplib2==0.2.0

# ── Testing ──────────────────────────────────────────────────────────────────
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0
locust>=2.20.0
